rest/requirements.txt
Django==1.9.1
Markdown==2.6.5
django-filter==0.12.0
djangorestframework==3.3.2
waitress==0.8.10
wsgiref==0.1.2
